Initial commit
This is the inital commit for Beacon as a standalone service. It pulls out the Beacon app from the overall purchasing suite and allows it to stand by itself for redeployment. Forthcoming features include:

+ A docker container for easier development/deployment
+ Generalization of templates away from using City of Pittsburgh for Everything
+ Better support for additional categories (using any spreadsheet with correct headers to create Categories)

As of this commit, doc references are broken from being ported over from the original purchasing suite. However, the app and tests should all pass as expected.