create extensible single-table inheritance for opp types
in order to support screendoor-based[1] submissions, we want to be able to support numerous different opportunity types. we use a single-table inheritance pattern (as supported by sqlalchemy[2]). each inherited class is responsible for the following:

   1) rendering submission instructions
   2) rendering a submission form if necessary
   3) serilaizing any additional data needed to render a submission screen
   4) validating any data needed for the submission screen

once a new subclass of the base Opportunity model is declared, it is automatically added to a Select dropdown on the new opportunity creation form. data is passed through, and the correct Opportunity subclass is constructed based on the data that comes in from the form, assuming valid data is passed.

fixes #5

[1] https://www.dobt.co/screendoor/
[2] http://docs.sqlalchemy.org/en/latest/orm/extensions/declarative/inheritance.html#single-table-inheritance