allow configurable help-blocks in opportunity classes
a new classmethod, get_help_block has now been added to the individual implemented opportunity subclasses. this classmethod, along with a new help-block rendering widget, attaches a 'data-help-block' attribute to each option in the opportunity type select field in the opportunity creation form. this text is then displayed as a help-block whenever that option is chosen.